Flyway config

+++ example/src/main/java/com/kgarbacki/customer/CustomerDao.java
@@ -10,15 +10,15 @@ public interface CustomerDao {
 
     List<Customer> selectAllCustomers();
 
-    Optional<Customer> selectCustomerById(Integer customerId);
+    Optional<Customer> selectCustomerById(Long customerId);
 
     void insertCustomer(Customer customer);
 
     boolean existsPersonWithEmail(String email);
 
-    boolean existsPersonWithId(Integer id);
+    boolean existsPersonWithId(Long id);
 
-    void deleteCustomerById(Integer customerId);
+    void deleteCustomerById(Long customerId);
 
     void updateCustomer(Customer update);
 }
