+++ example/src/main/java/com/kgarbacki/customer/CustomerService.java
@@ -23,7 +23,7 @@ public class CustomerService {
         return customerDao.selectAllCustomers();
     }
 
-    public Customer getCustomer(Integer id){
+    public Customer getCustomer(Long id){
         return customerDao.selectCustomerById(id)
                 .orElseThrow(() ->  new ResourceNotFoundException("Customer with Id: " + id + "does not exist"));
     }
@@ -45,7 +45,7 @@ public class CustomerService {
         customerDao.insertCustomer(newCustomer);
     }
 
-    public void deleteCustomerById(Integer customerId) {
+    public void deleteCustomerById(Long customerId) {
         if (!customerDao.existsPersonWithId(customerId)) {
             throw new ResourceNotFoundException(
                     "customer with id [%s] not found".formatted(customerId)
@@ -55,7 +55,7 @@ public class CustomerService {
         customerDao.deleteCustomerById(customerId);
     }
 
-    public void updateCustomer(Integer customerId,
+    public void updateCustomer(Long customerId,
                                CustomerUpdateRequest request) {
         Customer customer = getCustomer(customerId);
